requirements.txt

dill==0.3.5.1
from-root==1.0.2
notebook==7.0.0a7
boto3
mypy-boto3-s3
flask-cors
flask

matplotlib>=3.2.2
numpy>=1.18.5,<1.24.0
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
protobuf<4.21.3

tensorboard>=2.4.1
pandas>=1.1.4
seaborn>=0.11.0

ipython  
psutil  
thop  
ultralytics

-e .
